feat!: configurable autokeys, set buffer / window local options separately
## Details

Add `autokeys` configuration which defines the order in which keymaps
are constructed, this replaces the existing "start at a and go up"
logic, but allows for more values than ending at lowercase z, users can
modify it as they wish as well.

Breaking change around how options are set. In order to use the
`nvim_set_option_value` API options need to be separated based on
whether they are buffer local or window local. Default options have been
adjusted to this new mechanism. The `options` configuration has been
replaced with 2, `bo` for buffer local and `wo` for window local.